Initial Upload
Inital game system, and faction upload.  Points and force orgs complete
for Britain, USA, and Russia's main rule book org.  I need the books for
Afgantsy & W. German so it may be a couple of days for it.  I may work
on adding profiles the forces in the future, but not sure how to do it
and have it looking like a hot mess.